Remove unused method and move util to model
